requirements.txt
accelerate==1.7.0
bitsandbytes==0.45.5
ExifRead==3.3.1
imagesize==1.4.1
pillow==11.2.1
pillow-jxl-plugin~=1.3.4
pyparsing==3.2.1
PySide6==6.9.0
transformers==4.48.3

# Video support
opencv-python==4.10.0.84
# Optional experimental backend wrapper (requires native mpv runtime files)
python-mpv==1.0.8
# Optional experimental backend wrapper (requires native vlc runtime files)
python-vlc==3.0.21203

# CogAgent
timm==1.0.15

# CogVLM
einops==0.8.1
protobuf==6.31.0
sentencepiece==0.2.0
torchvision==0.21.0
xformers==0.0.29.post3

# InternLM-XComposer2
# gptqmodel==1.9.0
numpy==2.2.6

# WD Tagger
huggingface-hub==0.31.4
onnxruntime==1.22.0

# ============================================
# PyTorch & FlashAttention - DO NOT ADD HERE
# ============================================
# These are installed automatically by start_windows.bat / start_linux.sh
# The startup scripts detect your GPU's CUDA version and install the correct PyTorch build
# DO NOT manually add torch/flash-attn to this file - it will break CUDA compatibility
# ============================================

# YOLO
ultralytics>=8.3.100

# Spell checking and grammar checking
pyspellchecker==0.8.1
language-tool-python>=2.10

# Skin system
PyYAML>=6.0
